Improve MapView reimplementation
